Backend traits, renaming and adding a lock to lazy report
